+++ frontend/src/ts/constants/default-config.ts
@@ -105,4 +105,6 @@ const obj = {
   maxLineWidth: 0,
 } as Config;
 
-export default deepClone(obj);
+export function getDefaultConfig(): Config {
+  return deepClone(obj);
+}
