+++ frontend/src/ts/constants/default-snapshot.ts
@@ -1,5 +1,94 @@
+import {
+  ResultFilters,
+  User,
+  UserProfileDetails,
+  UserTag,
+} from "@monkeytype/contracts/schemas/users";
 import { deepClone } from "../utils/misc";
-import defaultConfig from "./default-config";
+import { getDefaultConfig } from "./default-config";
+import { Mode } from "@monkeytype/contracts/schemas/shared";
+import { Result } from "@monkeytype/contracts/schemas/results";
+import { Config } from "@monkeytype/contracts/schemas/configs";
+import {
+  ModifiableTestActivityCalendar,
+  TestActivityCalendar,
+} from "../elements/test-activity-calendar";
+import { Preset } from "@monkeytype/contracts/schemas/presets";
+
+export type SnapshotUserTag = UserTag & {
+  active?: boolean;
+  display: string;
+};
+
+export type SnapshotResult<M extends Mode> = Omit<
+  Result<M>,
+  | "_id"
+  | "bailedOut"
+  | "blindMode"
+  | "lazyMode"
+  | "difficulty"
+  | "funbox"
+  | "language"
+  | "numbers"
+  | "punctuation"
+  | "quoteLength"
+  | "restartCount"
+  | "incompleteTestSeconds"
+  | "afkDuration"
+  | "tags"
+> & {
+  _id: string;
+  bailedOut: boolean;
+  blindMode: boolean;
+  lazyMode: boolean;
+  difficulty: string;
+  funbox: string;
+  language: string;
+  numbers: boolean;
+  punctuation: boolean;
+  quoteLength: number;
+  restartCount: number;
+  incompleteTestSeconds: number;
+  afkDuration: number;
+  tags: string[];
+};
+
+export type Snapshot = Omit<
+  User,
+  | "timeTyping"
+  | "startedTests"
+  | "completedTests"
+  | "profileDetails"
+  | "streak"
+  | "resultFilterPresets"
+  | "tags"
+  | "xp"
+  | "testActivity"
+> & {
+  typingStats: {
+    timeTyping: number;
+    startedTests: number;
+    completedTests: number;
+  };
+  details?: UserProfileDetails;
+  inboxUnreadSize: number;
+  streak: number;
+  maxStreak: number;
+  filterPresets: ResultFilters[];
+  isPremium: boolean;
+  streakHourOffset?: number;
+  config: Config;
+  tags: SnapshotUserTag[];
+  presets: SnapshotPreset[];
+  results?: SnapshotResult<Mode>[];
+  xp: number;
+  testActivity?: ModifiableTestActivityCalendar;
+  testActivityByYear?: { [key: string]: TestActivityCalendar };
+};
+
+export type SnapshotPreset = Preset & {
+  display: string;
+};
 
 const defaultSnap = {
   results: undefined,
@@ -14,7 +103,7 @@ const defaultSnap = {
   email: "",
   uid: "",
   isPremium: false,
-  config: defaultConfig,
+  config: getDefaultConfig(),
   customThemes: [],
   presets: [],
   tags: [],
@@ -42,6 +131,8 @@ const defaultSnap = {
       60: { english: { count: 0, rank: 0 } },
     },
   },
-};
+} as Snapshot;
 
-export default deepClone(defaultSnap);
+export function getDefaultSnapshot(): Snapshot {
+  return deepClone(defaultSnap);
+}
