+++ frontend/src/ts/db.ts
@@ -1,7 +1,6 @@
 import Ape from "./ape";
 import * as Notifications from "./elements/notifications";
 import * as LoadingPage from "./pages/loading";
-import DefaultConfig from "./constants/default-config";
 import { isAuthenticated } from "./firebase";
 import * as ConnectionState from "./states/connection";
 import { lastElementFromArray } from "./utils/arrays";
@@ -13,14 +12,7 @@ import {
 } from "./elements/test-activity-calendar";
 import * as Loader from "./elements/loader";
 
-import {
-  Badge,
-  CustomTheme,
-  ResultFilters,
-  User,
-  UserProfileDetails,
-  UserTag,
-} from "@monkeytype/contracts/schemas/users";
+import { Badge, CustomTheme } from "@monkeytype/contracts/schemas/users";
 import { Config, Difficulty } from "@monkeytype/contracts/schemas/configs";
 import {
   Mode,
@@ -28,85 +20,15 @@ import {
   PersonalBest,
   PersonalBests,
 } from "@monkeytype/contracts/schemas/shared";
-import { Preset } from "@monkeytype/contracts/schemas/presets";
-import defaultSnapshot from "./constants/default-snapshot";
-import { Result } from "@monkeytype/contracts/schemas/results";
+import {
+  getDefaultSnapshot,
+  Snapshot,
+  SnapshotPreset,
+  SnapshotResult,
+  SnapshotUserTag,
+} from "./constants/default-snapshot";
+import { getDefaultConfig } from "./constants/default-config";
 import { FunboxMetadata } from "../../../packages/funbox/src/types";
-
-export type SnapshotUserTag = UserTag & {
-  active?: boolean;
-  display: string;
-};
-
-export type SnapshotResult<M extends Mode> = Omit<
-  Result<M>,
-  | "_id"
-  | "bailedOut"
-  | "blindMode"
-  | "lazyMode"
-  | "difficulty"
-  | "funbox"
-  | "language"
-  | "numbers"
-  | "punctuation"
-  | "quoteLength"
-  | "restartCount"
-  | "incompleteTestSeconds"
-  | "afkDuration"
-  | "tags"
-> & {
-  _id: string;
-  bailedOut: boolean;
-  blindMode: boolean;
-  lazyMode: boolean;
-  difficulty: string;
-  funbox: string;
-  language: string;
-  numbers: boolean;
-  punctuation: boolean;
-  quoteLength: number;
-  restartCount: number;
-  incompleteTestSeconds: number;
-  afkDuration: number;
-  tags: string[];
-};
-
-export type Snapshot = Omit<
-  User,
-  | "timeTyping"
-  | "startedTests"
-  | "completedTests"
-  | "profileDetails"
-  | "streak"
-  | "resultFilterPresets"
-  | "tags"
-  | "xp"
-  | "testActivity"
-> & {
-  typingStats: {
-    timeTyping: number;
-    startedTests: number;
-    completedTests: number;
-  };
-  details?: UserProfileDetails;
-  inboxUnreadSize: number;
-  streak: number;
-  maxStreak: number;
-  filterPresets: ResultFilters[];
-  isPremium: boolean;
-  streakHourOffset?: number;
-  config: Config;
-  tags: SnapshotUserTag[];
-  presets: SnapshotPreset[];
-  results?: SnapshotResult<Mode>[];
-  xp: number;
-  testActivity?: ModifiableTestActivityCalendar;
-  testActivityByYear?: { [key: string]: TestActivityCalendar };
-};
-
-export type SnapshotPreset = Preset & {
-  display: string;
-};
 
 let dbSnapshot: Snapshot | undefined;
 
@@ -147,7 +69,7 @@ export function setSnapshot(newSnapshot: Snapshot | undefined): void {
 
 export async function initSnapshot(): Promise<Snapshot | number | boolean> {
   //send api request with token that returns tags, presets, and data needed for snap
-  const snap = defaultSnapshot as Snapshot;
+  const snap = getDefaultSnapshot();
   try {
     if (!isAuthenticated()) return false;
     // if (ActivePage.get() === "loading") {
@@ -260,7 +182,7 @@ export async function initSnapshot(): Promise<Snapshot | number | boolean> {
     // LoadingPage.updateText("Downloading config...");
     if (configData === undefined || configData === null) {
       snap.config = {
-        ...DefaultConfig,
+        ...getDefaultConfig(),
       };
     } else {
       snap.config = migrateConfig(configData);
@@ -340,7 +262,7 @@ export async function initSnapshot(): Promise<Snapshot | number | boolean> {
     dbSnapshot = snap;
     return dbSnapshot;
   } catch (e) {
-    dbSnapshot = defaultSnapshot;
+    dbSnapshot = getDefaultSnapshot();
     throw e;
   }
 }
