+++ frontend/src/ts/modals/streak-hour-offset.ts
@@ -4,8 +4,9 @@ import * as Notifications from "../elements/notifications";
 import * as Loader from "../elements/loader";
 // import * as Settings from "../pages/settings";
 import * as ConnectionState from "../states/connection";
-import { getSnapshot, setSnapshot, Snapshot } from "../db";
+import { getSnapshot, setSnapshot } from "../db";
 import AnimatedModal from "../utils/animated-modal";
+import { Snapshot } from "../constants/default-snapshot";
 
 export function show(): void {
   if (!ConnectionState.get()) {
