+++ frontend/src/ts/utils/results.ts
@@ -4,6 +4,7 @@ import * as DB from "../db";
 import * as TestLogic from "../test/test-logic";
 import { deepClone } from "./misc";
 import { Mode } from "@monkeytype/contracts/schemas/shared";
+import { SnapshotResult } from "../constants/default-snapshot";
 
 export async function syncNotSignedInLastResult(uid: string): Promise<void> {
   const notSignedInLastResult = TestLogic.notSignedInLastResult;
@@ -26,7 +27,7 @@ export async function syncNotSignedInLastResult(uid: string): Promise<void> {
   // into a snapshot result - might not cuase issues but worth investigating
   const result = deepClone(
     notSignedInLastResult
-  ) as unknown as DB.SnapshotResult<Mode>;
+  ) as unknown as SnapshotResult<Mode>;
   result._id = response.body.data.insertedId;
   if (response.body.data.isPb) {
     result.isPb = true;
